requirements.txt
fastapi[all]
openai
python-dotenv
pydantic==1.*
langchain
bs4
tiktoken
langchain-experimental
faiss-cpu
sentence-transformers